Parser doesn't allocate anymore, add comments and reorganize code
